+++ src/libs/headers/drivers/tty.h
@@ -6,7 +6,7 @@
 #ifndef TTY_H
 #define TTY_H
 
-#include <stdint.h>
+#include "types.h"
 
 #define BUFFER_INPUT_SIZE 250
 #define MODE_CANON        0
@@ -18,10 +18,10 @@ struct tty_config {
 };
 
 // Writes count from buf to screen
-extern int ttyWrite(void* buf, size_t count);
+extern ssize_t ttyWrite(void* buf, size_t count);
 
 // Reads count from keyboard to buf
-extern int ttyRead(void* buf, size_t count);
+extern ssize_t ttyRead(void* buf, size_t count);
 
 // Controls special device specific functions
 extern int ttyIoctl(size_t request, ...);
